chore: respond to PR feedback
