+++ packages/utils/src/index.ts
@@ -8,7 +8,8 @@ export * from "./coreHelpers";
 export * from "./crypto";
 export * from "./email";
 export * from "./errors";
-export * from "./gqlConstants";
+export * from "./gql/queries";
+export * from "./gql/transforms";
 export * from "./handle";
 export * from "./language";
 export * from "./localStorage";
